api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
httpx==0.27.2
redis==5.0.4
SQLAlchemy>=2.0,<2.1
orjson==3.10.7
itsdangerous==2.2.0
psycopg2-binary==2.9.9
email-validator==2.2.0
